Corrections made

+++ project.py
@@ -21,7 +21,8 @@ from functools import wraps
 
 app = Flask(__name__)
 
-CLIENT_ID = json.loads(open('client_secrets.json', 'r').read())['web']['client_id']
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
 APPLICATION_NAME = "Restaurant Menu Application"
 
 engine = create_engine('sqlite:///restaurantappmenusource.db')
@@ -52,7 +53,8 @@ def showlogin():
 def gconnect():
     # validate state token
     if request.args.get('state')!= login_session['state']:
-        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response = make_response(
+            json.dumps('Invalid state parameter.'), 401)
         response.headers['Content-Type']= 'application-json'
         return response
     # Obtain authorization code
@@ -64,7 +66,8 @@ def gconnect():
         oauth_flow.redirect_uri = 'postmessage'
         credentials = oauth_flow.step2_exchange(code)
     except FlowExchangeError:
-        response = make_response(json.dumps('Failed to upgrade the authorization code'), 401)
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code'), 401)
         response.headers['Content-Type']= 'application-json'
         return response
 
@@ -174,24 +177,25 @@ def gdisconnect():
     print 'User name is: '
     print login_session['username']
     if access_token is None:
- 	print'Access Token is None'
-    	response=make_response(json.dumps('Current user not connected'), 401)
-    	response.headers['Content-Type']='application/json'
-    	return response
+      print'Access Token is None'
+      response=make_response(
+          json.dumps('Current user not connected'), 401)
+      response.headers['Content-Type']='application/json'
+      return response
     url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is'
     print result
     if result['status'] == '200':
-        response = make_response(json.dumps('Successfully disconnected.'), 200)
-    	response.headers['Content-Type'] = 'application/json'
-    	return response
+      response = make_response(json.dumps('Successfully disconnected.'), 200)
+      response.headers['Content-Type'] = 'application/json'
+      return response
     else:
-
-    	response = make_response(json.dumps('Failed to revoke token for given user.', 400))
-    	response.headers['Content-Type'] = 'application/json'
-    	return response
+      response = make_response(
+          json.dumps('Failed to revoke token for given user.', 400))
+      response.headers['Content-Type'] = 'application/json'
+      return response
 
 @app.route('/logout')
 def logout():
@@ -244,14 +248,18 @@ def showRestaurants():
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
 def newRestaurant():
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
-        newRestaurant = Restaurant(name = request.form['name'], user_id = login_session['user_id'])
-        session.add(newRestaurant)
-        session.commit()
-        flash('New Restaurant %s Successfully Created' % newRestaurant.name)
-        return redirect(url_for('showRestaurants'))
+      newRestaurant = Restaurant(
+         name = request.form['name'],
+         user_id = login_session['user_id'])
+      session.add(newRestaurant)
+      session.commit()
+      flash('New Restaurant %s Successfully Created' % newRestaurant.name)
+      return redirect(url_for('showRestaurants'))
     else:
-        return render_template('newRestaurant.html')
+      return render_template('newRestaurant.html')
 
 #Edit a restaurant
 @app.route('/restaurant/<int:restaurant_id>/edit/', methods = ['GET', 'POST'])
@@ -260,14 +268,16 @@ def editRestaurant(restaurant_id):
   if 'username' not in login_session:
     return redirect('/login')
   if editedRestaurant.user_id != login_session.get('user_id'):
-    return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
+    return "<script>function myFunction() {alert ('you are not authorized to \
+         edit this restaurant.please create your own restaurant \
+         in order to delete items');}</script><body onload='myFunction()''>"
   if request.method == 'POST':
-      if request.form['name']:
-        editedRestaurant.name = request.form['name']
-      session.add(editedRestaurant)
-      session.commit()
-      flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
-      return redirect(url_for('showRestaurants'))
+    if request.form['name']:
+      editedRestaurant.name = request.form['name']
+    session.add(editedRestaurant)
+    session.commit()
+    flash('Restaurant Successfully Edited %s' % editedRestaurant.name)
+    return redirect(url_for('showRestaurants'))
   else:
     return render_template('editRestaurant.html', restaurant = editedRestaurant)
 
@@ -275,10 +285,12 @@ def editRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods = ['GET','POST'])
 def deleteRestaurant(restaurant_id):
     restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if 'username' not in login_session:
+      return redirect('/login')
     if restaurantToDelete.user_id != login_session['user_id']:
-       return "<script>function myFunction() {alert('you are not authorized to \
-         delete this restaurant.please create your own restaurant to delete');}\
-         </script><body onLoad = 'myFunction()''>"
+      return "<script>function myFunction() {alert ('you are not authorized to \
+         delete this restaurant.please create your own restaurant \
+         in order to delete items');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
       session.delete(restaurantToDelete)
       flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -294,47 +306,50 @@ def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).first()
     creator = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-
     if 'username' not in login_session:
-       return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)
+      return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)
     else:
-       return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
+      return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
 
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
 def newMenuItem(restaurant_id):
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if 'username' not in login_session:
+    return redirect('/login')
   if request.method == 'POST':
-     newItem = MenuItem(name = request.form['name'], description = request.form['description'],
+    newItem = MenuItem(name = request.form['name'], description = request.form['description'],
                         price = request.form['price'], course = request.form['course'],
                         restaurant_id = restaurant_id, user_id=restaurant.user_id)
-     session.add(newItem)
-     session.commit()
-     flash('New Menu %s Item Successfully Created' % (newItem.name))
-     return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
+    session.add(newItem)
+    session.commit()
+    flash('New Menu %s Item Successfully Created' % (newItem.name))
+    return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
   else:
-      return render_template('newmenuitem.html', restaurant_id = restaurant_id)
+    return render_template('newmenuitem.html', restaurant_id = restaurant_id)
 
 #Edit a menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/<int:menu_id>/edit', methods=['GET','POST'])
 def editMenuItem(restaurant_id, menu_id):
     editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if 'username' not in login_session:
+      return redirect('/login')
     if login_session['user_id'] != restaurant.user_id:
-        return "<script>function myFunction() {alert('You are not authorized to \
-          edit menu items to this restaurant.Please create your own restaurant in \
-          order to edit items.');}</script><body onload='myFunction()''>"
+      return "<script>function myFunction() {alert ('you are not authorized to \
+         delete menu items to this restaurant.please create your own restaurant \
+         in order to edit items');}</script><body onload='myFunction()''>"
 
     if request.method == 'POST':
         if request.form['name']:
-            editedItem.name = request.form['name']
+          editedItem.name = request.form['name']
         if request.form['description']:
-            editedItem.description = request.form['description']
+          editedItem.description = request.form['description']
         if request.form['price']:
-            editedItem.price = request.form['price']
+          editedItem.price = request.form['price']
         if request.form['course']:
-            editedItem.course = request.form['course']
+          editedItem.course = request.form['course']
         session.add(editedItem)
         session.commit()
         flash('Menu Item Successfully Edited')
@@ -350,20 +365,22 @@ def editMenuItem(restaurant_id, menu_id):
 def deleteMenuItem(restaurant_id,menu_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one()
+    if 'username' not in login_session:
+      return redirect('/login')
     if login_session['user_id'] != restaurant.user_id:
-       return "<script>function myFunction() {alert ('you are not authorized to \
+      return "<script>function myFunction() {alert ('you are not authorized to \
          delete menu items to this restaurant.please create your own restaurant \
          in order to delete items');}</script><body onload='myFunction()''>"
     if request.method == 'POST':
-        session.delete(itemToDelete)
-        session.commit()
-        flash('Menu Item Successfully Deleted')
-        return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
+      session.delete(itemToDelete)
+      session.commit()
+      flash('Menu Item Successfully Deleted')
+      return redirect(url_for('showRestaurants', restaurant_id = restaurant_id))
     else:
-        return render_template('deleteMenuItem.html', item = itemToDelete)
+      return render_template('deleteMenuItem.html', item = itemToDelete)
 
 
 if __name__ == '__main__':
-   app.secret_key = 'super_secret_key'
-   app.debug = True
-   app.run(host = '0.0.0.0', port = 8000)
+  app.secret_key = 'super_secret_key'
+  app.debug = True
+  app.run(host = '0.0.0.0', port = 8000)
